Рефакторинг. Добавляет компоненты, для оптимизации перерисовок
