--- src/mocks/offers.ts
@@ -1,699 +1,0 @@
-import { Offer, OfferCardType } from '../types/index';
-
-export const OFFERS: Offer[] = [
-  {
-    id: '8a1bd7dd-ef78-43f6-859e-0d3c83413ef3',
-    title: 'Beautiful & luxurious studio at great location',
-    type: 'apartment',
-    price: 120,
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.35514938496378,
-        longitude: 4.673877537499948,
-        zoom: 8
-      }},
-    location: {
-      latitude: 52.3909553943508,
-      longitude: 4.85309666406198,
-      zoom: 8
-    },
-    isFavorite: true,
-    isPremium: true,
-    rating: 2,
-    description: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    bedrooms: 3,
-    goods: [
-      'Heating'
-    ],
-    host: {
-      name: 'Oliver Conner',
-      avatarUrl: 'img/avatar-max.jpg',
-      isPro: false
-    },
-    images: [
-      'img/apartment-01.jpg'
-    ],
-    maxAdults: 4
-  },
-  {
-    id: 'cc3a1b0f-e69f-464c-9b97-fef3f29d32cf',
-    title: 'Beautiful & luxurious studio at great location',
-    type: 'room',
-    price: 80,
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.35514938496378,
-        longitude: 4.673877537499948,
-        zoom: 8
-      }},
-    location: {
-      latitude: 52.3609553943508,
-      longitude: 4.85309666406198,
-      zoom: 8
-    },
-    isFavorite: true,
-    isPremium: true,
-    rating: 3,
-    description: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    bedrooms: 3,
-    goods: [
-      'Heating'
-    ],
-    host: {
-      name: 'Oliver Conner',
-      avatarUrl: 'img/avatar-max.jpg',
-      isPro: false
-    },
-    images: [
-      'img/apartment-02.jpg'
-    ],
-    maxAdults: 4
-  },
-  {
-    id: '23e1063d-4212-4c09-859b-4b5a819fe173',
-    title: 'Beautiful & luxurious studio at great location',
-    type: 'house',
-    price: 90,
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.35514938496378,
-        longitude: 4.673877537499948,
-        zoom: 8
-      }},
-    location: {
-      latitude: 52.3909553943508,
-      longitude: 4.929309666406198,
-      zoom: 8
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 5,
-    description: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    bedrooms: 3,
-    goods: [
-      'Heating'
-    ],
-    host: {
-      name: 'Oliver Conner',
-      avatarUrl: 'img/avatar-max.jpg',
-      isPro: false
-    },
-    images: [
-      'img/apartment-03.jpg'
-    ],
-    maxAdults: 4
-  },
-  {
-    id: 'de0f6ed4-2ef4-4820-a86e-08378b0a211b',
-    title: 'Beautiful & luxurious studio at great location',
-    type: 'hotel',
-    price: 30,
-    city: {
-      name: 'Amsterdam',
-      location: {
-        latitude: 52.35514938496378,
-        longitude: 4.673877537499948,
-        zoom: 8
-      }},
-    location: {
-      latitude: 52.3809553943508,
-      longitude: 4.939309666406198,
-      zoom: 8
-    },
-    isFavorite: false,
-    isPremium: false,
-    rating: 4,
-    description: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    bedrooms: 3,
-    goods: [
-      'Heating'
-    ],
-    host: {
-      name: 'Oliver Conner',
-      avatarUrl: 'img/avatar-max.jpg',
-      isPro: false
-    },
-    images: [
-      'img/room.jpg'
-    ],
-    maxAdults: 4
-  },
-];
-
-export const OFFER_CARDS: OfferCardType[] = [
-  {
-    'id': '8a1bd7dd-ef78-43f6-859e-0d3c83413ef3',
-    'title': 'Penthouse, 4-5 rooms + 5 balconies',
-    'type': 'room',
-    'price': 121,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/14.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.868610000000004,
-      'longitude': 2.342499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 1.7
-  },
-  {
-    'id': 'cc3a1b0f-e69f-464c-9b97-fef3f29d32cf',
-    'title': 'Tile House',
-    'type': 'apartment',
-    'price': 313,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/17.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.858610000000006,
-      'longitude': 2.330499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.5
-  },
-  {
-    'id': '23e1063d-4212-4c09-859b-4b5a819fe173',
-    'title': 'The house among olive ',
-    'type': 'apartment',
-    'price': 314,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/15.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.834610000000005,
-      'longitude': 2.335499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2
-  },
-  {
-    'id': 'de0f6ed4-2ef4-4820-a86e-08378b0a211b',
-    'title': 'Canal View Prinsengracht',
-    'type': 'room',
-    'price': 200,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/12.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.85761,
-      'longitude': 2.358499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.6
-  },
-  {
-    'id': '6a12627e-0caf-4df7-8532-a6c8925e2017',
-    'title': 'House in countryside',
-    'type': 'room',
-    'price': 293,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/6.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.950361,
-      'longitude': 6.961974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.8
-  },
-  {
-    'id': '6275ef9d-f679-4137-bb30-7b7916604516',
-    'title': 'The house among olive ',
-    'type': 'hotel',
-    'price': 170,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/20.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.932361,
-      'longitude': 6.937974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.4
-  },
-  {
-    'id': '81f0392f-b9f8-4d2d-b6d7-1839d38d72c1',
-    'title': 'Beautiful & luxurious apartment at great location',
-    'type': 'room',
-    'price': 146,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/11.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.934361,
-      'longitude': 6.943974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.3
-  },
-  {
-    'id': '56716286-3043-4855-9b5a-a9bedae2cff2',
-    'title': 'Canal View Prinsengracht',
-    'type': 'hotel',
-    'price': 357,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/10.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.947361,
-      'longitude': 6.9799739999999995,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.1
-  },
-  {
-    'id': '0538c5da-533d-4ccf-94d2-dd119a847a78',
-    'title': 'House in countryside',
-    'type': 'room',
-    'price': 258,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/18.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.854557,
-      'longitude': 4.364697,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.4
-  },
-  {
-    'id': '73e5bba3-8c20-49a2-a3b4-44a290d76f36',
-    'title': 'Beautiful & luxurious apartment at great location',
-    'type': 'hotel',
-    'price': 323,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/18.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.867557,
-      'longitude': 4.371696999999999,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.6
-  },
-  {
-    'id': '2a9fdff6-a670-4ae5-bae6-877a8ccab8c6',
-    'title': 'The Joshua Tree House',
-    'type': 'room',
-    'price': 111,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/1.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.827557,
-      'longitude': 4.336697,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.5
-  },
-  {
-    'id': '27c7b012-16fb-4331-9b2f-e7c44a8e3208',
-    'title': 'The Joshua Tree House',
-    'type': 'hotel',
-    'price': 399,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/3.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.833557,
-      'longitude': 4.374696999999999,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.5
-  },
-  {
-    'id': '4f424408-31d4-44dd-80ae-c65819aa6bd0',
-    'title': 'Tile House',
-    'type': 'hotel',
-    'price': 115,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/9.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.837557,
-      'longitude': 4.339697,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.5
-  },
-  {
-    'id': 'd6e8b1eb-f59d-4b15-abea-b07e22359ac7',
-    'title': 'Amazing and Extremely Central Flat',
-    'type': 'hotel',
-    'price': 275,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/3.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.849557,
-      'longitude': 4.374696999999999,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.3
-  },
-  {
-    'id': 'b97bc5bb-1820-46fb-9907-271fadcb641e',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'house',
-    'price': 641,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/18.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.852557,
-      'longitude': 4.3376969999999995,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.6
-  },
-  {
-    'id': '1051e923-2036-47bf-ba02-d402926b4e1b',
-    'title': 'Tile House',
-    'type': 'apartment',
-    'price': 263,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/3.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.36554,
-      'longitude': 4.911976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.1
-  },
-  {
-    'id': '5aaf420c-94cb-4f51-8a6f-0c2204ebb14c',
-    'title': 'Canal View Prinsengracht',
-    'type': 'apartment',
-    'price': 249,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/19.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.385540000000006,
-      'longitude': 4.902976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.2
-  },
-  {
-    'id': 'b3305db6-ef61-48f3-bb4b-b67301f3191d',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'house',
-    'price': 857,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/1.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.397540000000006,
-      'longitude': 4.9099759999999995,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.2
-  },
-  {
-    'id': 'c1e65541-d4e4-4637-8f69-ffd6d0f91827',
-    'title': 'The Pondhouse - A Magical Place',
-    'type': 'house',
-    'price': 355,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/2.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.528341000000005,
-      'longitude': 10.018654000000002,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.5
-  },
-  {
-    'id': 'cc2d3cec-0b73-46cd-bee7-41d52eb2b84f',
-    'title': 'Tile House',
-    'type': 'house',
-    'price': 777,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/16.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.538341,
-      'longitude': 9.976654000000002,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.9
-  },
-  {
-    'id': 'a2124d2a-c2b6-4b45-a16b-897c6b2058d8',
-    'title': 'House in countryside',
-    'type': 'house',
-    'price': 532,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/9.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.563341,
-      'longitude': 10.019654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.4
-  },
-  {
-    'id': '849fb189-0b15-4164-8e3b-535b1c0f6284',
-    'title': 'Perfectly located Castro',
-    'type': 'room',
-    'price': 253,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/9.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.236402000000005,
-      'longitude': 6.784314,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2
-  },
-  {
-    'id': '09e0480f-9f4e-422a-a47e-df5b69d121bf',
-    'title': 'Tile House',
-    'type': 'room',
-    'price': 147,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/20.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.210402,
-      'longitude': 6.798314,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.7
-  },
-  {
-    'id': 'a3cc1106-cfce-4a1c-be99-f0e36c9e3bb2',
-    'title': 'Amazing and Extremely Central Flat',
-    'type': 'hotel',
-    'price': 451,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/7.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.211402,
-      'longitude': 6.756314000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.5
-  }
-];
